+++ Application/Table.hpp
@@ -21,29 +21,29 @@ namespace Game {
 
     struct TableLimitException : public std::logic_error
     {
-        TableLimitException(intmax_t currentBetAmount, intmax_t currentCosts, intmax_t tableLimit);
+        TableLimitException(int currentBetAmount, int currentCosts, int tableLimit);
         virtual char const * what() const throw();
-        intmax_t betAmount_;
-        intmax_t tableCost_;
-        intmax_t tableLimit_;
+        int betAmount_;
+        int tableCost_;
+        int tableLimit_;
 
     };
 
     class Table {
     public:
-        Table (intmax_t limit=1000);
+        Table (int limit=1000);
         virtual ~Table ();
         void PlaceBet(Bet& bet, Hand& hand);
         bool IsValid(Bet& bet) const;
         void AddPlayer();
         string ToString() const;
     private:
-        intmax_t limit_;
+        int limit_;
         vector<Player*> _players;
         vector<shared_ptr<Bet>> bets_;
-        intmax_t Cost() const;
+        int Cost() const;
     };
     
-} /* game */
+} /* Casino */
 
 #endif /* end of include guard: _TABLE */
